Generate text at build time
Generate text when building instead of hard-coding the text in a bitmap
image.  This uses the text2chr utility to build the variable width text
image from a given font image.

There are four images generated, one for each line of  text, two lines
per screen.  None of these can be blank, or things start to break.  The
text to be used is in the Makefile as text1, text2, text3, and text4.